Crate vectoreyes

VectorEyes is a (almost entirely) safe wrapper library around vectorized operations.

Backends

VectorEyes chooses what backend to execute vector operations with at compile-time.

AVX2

CPUs that support the AVX, AVX2, SSE4.1, AES, SSE4.2, and PCLMULQDQ features will use the AVX2 backend.

In addition, we have embedded specific latency numbers for:

skylake

skylake-avx512

cascadelake

znver1

znver2

znver3

As a result, vectoreyes will be more efficient on these platforms. You can add specific latency numbers for more targets in avx2.py.

Scalar

At the moment, this is the only alternative to the AVX2 backend. It is not particularly optimized.

Cargo Configuration

Native CPU Setup

Compile on the machine that you’ll be running your code on, and add the following to your .cargo/config file:

[build]
rustflags = ["-C", "target-cpu=native", "--cfg=vectoreyes-target-cpu-native"]
rustdocflags = ["-C", "target-cpu=native", "--cfg=vectoreyes-target-cpu-native"]

Specific CPU Selection

If you want to compile for some specific CPU

[build]
rustflags = ["-C", "target-cpu=TARGET", "--cfg=vectoreyes-target-cpu=\"TARGET\""]
rustdocflags = ["-C", "target-cpu=TARGET", "--cfg=vectoreyes-target-cpu=\"TARGET\""]

Maximal Compatibility

If you do not put any of the above in your .cargo/config file, vectoreyes will always use its scalar backend, which does not use vector instructions.

NOTE: many functions are currently missing from this library. Please consult the Intel documentation to see if a non-implemented intrinsic would more directly accomplish your goal, and we can add it!

Modules

array_utils

Perform manually unrolled operations on arrays.

Structs

Aes128

Aes128EncryptOnly

Aes256

Aes256EncryptOnly

I8x16

[i8; 16] as a vector.

I8x32

[i8; 32] as a vector.

I16x8

[i16; 8] as a vector.

I16x16

[i16; 16] as a vector.

I32x4

[i32; 4] as a vector.

I32x8

[i32; 8] as a vector.

I64x2

[i64; 2] as a vector.

I64x4

[i64; 4] as a vector.

U8x16

[u8; 16] as a vector.

U8x32

[u8; 32] as a vector.

U16x8

[u16; 8] as a vector.

U16x16

[u16; 16] as a vector.

U32x4

[u32; 4] as a vector.

U32x8

[u32; 8] as a vector.

U64x2

[u64; 2] as a vector.

U64x4

[u64; 4] as a vector.

Enums

MicroArchitecture

VectorBackend

Constants

VECTOR_BACKEND

The backend that is used to evaluate vector operations.

Traits

AesBlockCipher

AesBlockCipherDecrypt

ExtendingCast

Lossily cast a vector by {zero,sign}-extending its values.

HasVector

A utility trait you probably won’t need to use. See Simd.

Scalar

A scalar that can live in the lane of a vector.

SimdBase

A vector equivalent to [T; Self::Lanes].

SimdBase4x

A vector containing 4 lanes.

SimdBase4x64

A vector containing 4 64-bit values.

SimdBase8

A vector containing 8-bit values.

SimdBase8x

A vector containing 8 lanes.

SimdBase32

A vector containing 32-bit values.

SimdBase64

A vector containing 64-bit values.

SimdBaseGatherable

Type Definitions

Simd

An alternative way of naming SIMD types.