Enum vectoreyes::array_utils::ArrayUnrolledOps

pub enum ArrayUnrolledOps {}

A type which can be used to require that unrolled operations exist for a given array size.

Example

use vectoreyes::array_utils::*;
pub fn make_fun_array<const N: usize>() -> [usize; N]
    where ArrayUnrolledOps: UnrollableArraySize<N>
{
    <[usize; N]>::array_generate(|i| i + 10)
}
assert_eq!(make_fun_array::<4>(), [10, 11, 12, 13]);