Trait vectoreyes::ExtendingCast

pub trait ExtendingCast<T: SimdBase>: SimdBase {
    fn extending_cast_from(t: T) -> Self;
}

Lossily cast a vector by {zero,sign}-extending its values.

Required Methods

fn extending_cast_from(t: T) -> Self

Cast from one vector to another by sign or zero exending the values from the source until it fills the destination.

This operation is neccessarily lossy. The lowest-index values in t are kept. Other values are discarded.