Trait vectoreyes::Scalar

pub trait Scalar: 'static + Debug + PrimInt + WrappingAdd + WrappingSub + WrappingMul + ConstantTimeEq + ConditionallySelectable {
    type Signed: Scalar;
    type Unsigned: Scalar;
    type SameSign8: Scalar<Signed = i8, Unsigned = u8>;
    type SameSign16: Scalar<Signed = i16, Unsigned = u16>;
    type SameSign32: Scalar<Signed = i32, Unsigned = u32>;
    type SameSign64: Scalar<Signed = i64, Unsigned = u64>;
}

A scalar that can live in the lane of a vector.

Required Associated Types

type Signed: Scalar

A scalar of the same width as this scalar, but signed.

type Unsigned: Scalar

A scalar of the same width as this scalar, but unsigned.

type SameSign8: Scalar<Signed = i8, Unsigned = u8>

A scalar of the same sign as this scalar, but with width 8.

type SameSign16: Scalar<Signed = i16, Unsigned = u16>

A scalar of the same sign as this scalar, but with width 16.

type SameSign32: Scalar<Signed = i32, Unsigned = u32>

A scalar of the same sign as this scalar, but with width 32.

type SameSign64: Scalar<Signed = i64, Unsigned = u64>

A scalar of the same sign as this scalar, but with width 64.