Trait vectoreyes::SimdBaseGatherable

pub trait SimdBaseGatherable<IV: SimdBase>: SimdBase {
    unsafe fn gather(base: *const Self::Scalar, indices: IV) -> Self;
    unsafe fn gather_masked(
        base: *const Self::Scalar,
        indices: IV,
        mask: Self,
        src: Self
    ) -> Self;
}

Required Methods

unsafe fn gather(base: *const Self::Scalar, indices: IV) -> Self

Construct a vector by accessing values at base + indices[i]

unsafe fn gather_masked(
base: *const Self::Scalar,
indices: IV,
mask: Self,
src: Self
) -> Self

Construct a vector by accessing values at base + indices[i], only if the mask is set.